requirements.txt
streamlit
scikit-learn
pandas
joblib
matplotlibss